v1.4.0: Bug fixes, optimisation and new features
First thank you to Matt Schmitz (https://github.com/Matt-Schmitz) for identifying a major inefficiency in the overlap checking and providing a detailed fix (See - https://github.com/NickJD/ORForise/issues/11#issue-16045008540). This has provided a significant speedup for comparing annotations (Annotation-Compare and Aggregate-Compare).

GFF_Adder is currently undergoing a major rewrite and should be used with that in mind. The user options and general utility of the tool will change significantly over the next few months.
There was a bug in GFF_Adder which would sometimes mean that overlapping annotations would not be filtered out correctly.
GFF_Adder now has additional functionality which will be documented in the near future.